use float32

+++ views_pipeline_core/data/utils.py
@@ -13,14 +13,14 @@ def ensure_float64(df):
         and convert the DataFrame to use np.float64 for all its numeric columns.
         """
         non_float64_cols = df.select_dtypes(include=['number']).columns[
-            df.select_dtypes(include=['number']).dtypes != np.float64]
+            df.select_dtypes(include=['number']).dtypes != np.float32]
 
         if len(non_float64_cols) > 0:
             logger.warning(
                 f"DataFrame contains non-np.float64 numeric columns. Converting the following columns: {', '.join(non_float64_cols)}")
 
             for col in non_float64_cols:
-                df[col] = df[col].astype(np.float64)
+                df[col] = df[col].astype(np.float32)
         return df
 
 def download_json(url):
